requirements.txt
requests==2.28.2
pytest==7.2.1
pytest-html==3.2.0
allure-pytest==2.12.0
pathlib2==2.3.7.post1